Use infobip-jackson-extension
So it turns out the reason why Jackson deserialization sans-annotations was failing for FileDocumentProvider was because it has a constructor with exactly 1 field. See https://github.com/FasterXML/jackson-databind/issues/1498. This issue is supposed to have been fixed back in Jackson 2.12, but for some reason it's still broken for me in 2.18.2. Using the third-party module mentioned in the GitHub issue works.